Refactored client to use shared functions
